+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+* Autosave ([#144](https://github.com/i-VRESSE/workflow-builder/pull/144))
+* Improved handling of groups in ui schema
+* Nodes with all its properties in a single group is expanded by default
+
+### Fixed
+
+* Hide the exit module ([#153](https://github.com/i-VRESSE/workflow-builder/issues/153))
+
 ## @i-vresse/wb-core 2.0.1 - 2024-03-25
 
 ### Changed